% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_elements.R
\name{extract_elements}
\alias{extract_elements}
\title{Extract elements from row slot of gDist object}
\usage{
extract_elements(row)
}
\arguments{
\item{row}{gDist$rows[\link{ind}]}
}
\value{
data.frame
}
\description{
Extract elements from row slot of gDist object
}
