% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distances.R
\name{get_distances}
\alias{get_distances}
\title{Use the Google Distance Matrix API}
\usage{
get_distances(
  origins,
  destinations,
  departure_time = "now",
  .api_key,
  .output_format = "json"
)
}
\arguments{
\item{origins}{character vector. Origin addresses. These are compared to every value in
destinations.}

\item{destinations}{character vector. Destination addresses. These are compared to
every value in origins}

\item{departure_time}{a datetime (POSIXct) or 'now'. The desired time of departure.}

\item{.api_key}{google cloud API key. This must be setup and enabled to use the Google distance
matrix API.}

\item{.output_format}{character. The return format from the API call. Either 'json' or 'xml'}
}
\description{
The Distance Matrix API is a service that provides travel distance and time
for a matrix of origins and destinations. The API returns information based
on the recommended route between start and end points, as calculated by the
Google Maps API, and consists of rows containing duration and distance values
for each pair.
}
\details{
https://maps.googleapis.com/maps/api/distancematrix/outputFormat?parameters
}
