% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_distances.R
\name{wrangle_distances}
\alias{wrangle_distances}
\title{Convert returned Google Distance Matrix results into a data.frame}
\usage{
wrangle_distances(got_distances)
}
\arguments{
\item{got_distances}{object from get_distances()}
}
\description{
Convert returned Google Distance Matrix results into a data.frame
}
